IDEMPIERE-5991:improve file download by append file list inside process result panel
